SDCE-461. Replace ratelimit library with Apache 2 license compatible one
Reviewed-on: https://review.streamsets.net/c/19930
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
(cherry picked from commit c658927221dcf29655ebc1e928856ed474ec1707)
Reviewed-on: https://review.streamsets.net/c/19932

+++ stages/origins/spooler/whole_file.go
@@ -32,7 +32,7 @@ func (s *SpoolDirSource) produceWholeFileRecord(batchMaker api.BatchMaker) {
 	fileRef := wholefilerecord.NewLocalFileRef(
 		fInfo.getFullPath(),
 		cast.ToInt64(s.Conf.DataFormatConfig.WholeFileMaxObjectLen),
-		cast.ToFloat64(s.Conf.DataFormatConfig.RateLimit),
+		cast.ToInt(s.Conf.DataFormatConfig.RateLimit),
 	)
 
 	recordReaderFactory := s.Conf.DataFormatConfig.RecordReaderFactory
